refactor(core): migrate storage and debug log owners
