+++ src/crates/services/services-core/src/token_usage/mod.rs
@@ -1,5 +1,7 @@
+mod service;
 pub mod types;
 
+pub use service::TokenUsageService;
 pub use types::{
     ModelTokenStats, SessionTokenStats, TimeRange, TokenUsageQuery, TokenUsageRecord,
     TokenUsageSummary,
